+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Support for system keychain for storing passwords
 - Fix dialog windows staying open even after game launch (eframe bug, switched to popout)
 - Fix manifest url config option being ignored
+- Add config validation